feat(ui): rework the routes tab and extract HttpRoutesPanel
Match the mockup and the frontend proxies tab: put the search box, the
node combo and a clear filters link on one row, drop the search box the
data table owned, and make the create button primary. Filtering to
nothing now shows a no results state instead of an empty table.

The node filter notification above the tabs is gone, the clear filters
link replaces it.

Move the tab body and its three modals to HttpRoutesPanel, mirroring
HttpFrontendProxiesPanel. The view keeps the query params, the tab
selector and the single list-installed-modules chain that feeds both
panels; each panel reloads from a traefikInstances watcher.

The node filter stays a query param, so the panel takes it as a prop and
mirrors it locally: repainting the combo box needs a synchronous write
that a round trip through the URL cannot guarantee.
