+++ core/ui/public/i18n/en/translation.json
@@ -788,16 +788,13 @@
         "name": "Name",
         "type": "Type",
         "node": "Node",
-        "search_route": "Search route",
+        "search_route": "Filter routes",
         "no_http_route": "No HTTP route",
         "no_http_route_description": "There is no route configured",
         "create_route": "Create route",
         "host": "Host",
         "path": "Path",
         "host_and_path": "Host and path",
-        "routes_filtered": "Routes filtered",
-        "routes_filtered_description": "Showing only routes for {node}",
-        "show_all": "Show all",
         "http_route_details": "HTTP route details",
         "url": "URL",
         "strip_prefix": "Strip URL path prefix",
